+++ test/DeclarativePipelineTest.groovy
@@ -19,18 +19,6 @@ class DeclarativePipelineTest {
     }
 
     @Nested
-    public class WithNodeLabel {
-        @Test
-        void isFluent() {
-            def pipeline = new DeclarativePipeline(new MockWorkflowScript())
-
-            def result = pipeline.withNodeLabel('foo')
-
-            assertThat(result, equalTo(pipeline))
-        }
-    }
-
-    @Nested
     public class StartsWith {
         @Test
         void isFluent() {
@@ -59,7 +47,7 @@ class DeclarativePipelineTest {
         @Test
         void buildsThePipelineUsingTheAppropriateDeclarativeTemplate() {
             def wasCalled = false
-            def pipelineTemplate = { stages, label -> wasCalled = true }
+            def pipelineTemplate = { stages -> wasCalled = true }
             def pipeline = spy(new DeclarativePipeline(new MockWorkflowScript()))
             doReturn(pipelineTemplate).when(pipeline).getPipelineTemplate(any(List))
 
@@ -72,7 +60,7 @@ class DeclarativePipelineTest {
         void correctlyOrdersTheStages() {
             def expectedStages = [mock(Stage), mock(Stage), mock(Stage)]
             def actualStages = null
-            def pipelineTemplate = { stages, label -> actualStages = stages }
+            def pipelineTemplate = { stages -> actualStages = stages }
             def pipeline = spy(new DeclarativePipeline(new MockWorkflowScript()))
             doReturn(pipelineTemplate).when(pipeline).getPipelineTemplate(any(List))