+++ vars/Pipeline1Stage.groovy
@@ -1,4 +1,4 @@
-def call(args, label) {
+def call(args) {
     pipeline {
         agent none
         options { preserveStashes() }
@@ -7,11 +7,9 @@ def call(args, label) {
             stage('1') {
                 steps {
                     script {
-                        node(label) {
-                            def configuration = ((Stage)args.getAt(0)).pipelineConfiguration()
-                            configuration.delegate = this
-                            configuration()
-                        }
+                        def configuration = ((Stage)args.getAt(0)).pipelineConfiguration()
+                        configuration.delegate = this
+                        configuration()
                     }
                 }
             }